+++ internal/compat/css_table.go
@@ -1,5 +1,9 @@
 package compat
 
+import (
+	"github.com/evanw/esbuild/internal/css_ast"
+)
+
 type CSSFeature uint8
 
 const (
@@ -105,6 +109,164 @@ func UnsupportedCSSFeatures(constraints map[Engine][]int) (unsupported CSSFeatur
 			if versionRanges, ok := engines[engine]; !ok || !isVersionSupported(versionRanges, version) {
 				unsupported |= feature
 			}
+		}
+	}
+	return
+}
+
+type CSSPrefix uint8
+
+const (
+	WebkitPrefix CSSPrefix = 1 << iota
+	MozPrefix
+	MsPrefix
+	OPrefix
+
+	NoPrefix CSSPrefix = 0
+)
+
+type prefixData struct {
+	// Note: In some cases, earlier versions did not require a prefix but later
+	// ones do. This is the case for Microsoft Edge for example, which switched
+	// the underlying browser engine from a custom one to the one from Chrome.
+	// However, we assume that users specifying a browser version for CSS mean
+	// "works in this version or newer", so we still add a prefix when a target
+	// is an old Edge version.
+	withoutPrefix v
+	prefix        CSSPrefix
+}
+
+var cssMaskPrefixTable = map[Engine]prefixData{
+	Chrome: {prefix: WebkitPrefix},
+	Edge:   {prefix: WebkitPrefix},
+	IOS:    {prefix: WebkitPrefix, withoutPrefix: v{15, 4, 0}},
+	Opera:  {prefix: WebkitPrefix},
+	Safari: {prefix: WebkitPrefix, withoutPrefix: v{15, 4, 0}},
+}
+
+var cssPrefixTable = map[css_ast.D]map[Engine]prefixData{
+	// https://caniuse.com/css-appearance
+	css_ast.DAppearance: {
+		Chrome:  {prefix: WebkitPrefix, withoutPrefix: v{84, 0, 0}},
+		Edge:    {prefix: WebkitPrefix, withoutPrefix: v{84, 0, 0}},
+		Firefox: {prefix: MozPrefix, withoutPrefix: v{80, 4, 0}},
+		IOS:     {prefix: WebkitPrefix, withoutPrefix: v{15, 4, 0}},
+		Opera:   {prefix: WebkitPrefix, withoutPrefix: v{73, 4, 0}},
+		Safari:  {prefix: WebkitPrefix, withoutPrefix: v{15, 4, 0}},
+	},
+
+	// https://caniuse.com/css-backdrop-filter
+	css_ast.DBackdropFilter: {
+		IOS:    {prefix: WebkitPrefix},
+		Safari: {prefix: WebkitPrefix},
+	},
+
+	// https://caniuse.com/background-clip-text (Note: only for "background-clip: text")
+	css_ast.DBackgroundClip: {
+		Chrome: {prefix: WebkitPrefix},
+		Edge:   {prefix: WebkitPrefix},
+		IOS:    {prefix: WebkitPrefix, withoutPrefix: v{14, 0, 0}},
+		Opera:  {prefix: WebkitPrefix},
+		Safari: {prefix: WebkitPrefix, withoutPrefix: v{14, 0, 0}},
+	},
+
+	// https://caniuse.com/css-clip-path
+	css_ast.DClipPath: {
+		Chrome: {prefix: WebkitPrefix, withoutPrefix: v{55, 0, 0}},
+		IOS:    {prefix: WebkitPrefix, withoutPrefix: v{13, 0, 0}},
+		Opera:  {prefix: WebkitPrefix, withoutPrefix: v{42, 0, 0}},
+		Safari: {prefix: WebkitPrefix, withoutPrefix: v{13, 1, 0}},
+	},
+
+	// https://caniuse.com/font-kerning
+	css_ast.DFontKerning: {
+		Chrome: {prefix: WebkitPrefix, withoutPrefix: v{33, 0, 0}},
+		IOS:    {prefix: WebkitPrefix, withoutPrefix: v{12, 0, 0}},
+		Opera:  {prefix: WebkitPrefix, withoutPrefix: v{20, 0, 0}},
+		Safari: {prefix: WebkitPrefix, withoutPrefix: v{9, 1, 0}},
+	},
+
+	// https://caniuse.com/css-hyphens
+	css_ast.DHyphens: {
+		Edge:    {prefix: MsPrefix, withoutPrefix: v{79, 0, 0}},
+		Firefox: {prefix: MozPrefix, withoutPrefix: v{43, 0, 0}},
+		IE:      {prefix: MsPrefix},
+		IOS:     {prefix: WebkitPrefix},
+		Safari:  {prefix: WebkitPrefix},
+	},
+
+	// https://caniuse.com/css-initial-letter
+	css_ast.DInitialLetter: {
+		IOS:    {prefix: WebkitPrefix},
+		Safari: {prefix: WebkitPrefix},
+	},
+
+	css_ast.DMaskImage:    cssMaskPrefixTable, // https://caniuse.com/mdn-css_properties_mask-image
+	css_ast.DMaskOrigin:   cssMaskPrefixTable, // https://caniuse.com/mdn-css_properties_mask-origin
+	css_ast.DMaskPosition: cssMaskPrefixTable, // https://caniuse.com/mdn-css_properties_mask-position
+	css_ast.DMaskRepeat:   cssMaskPrefixTable, // https://caniuse.com/mdn-css_properties_mask-repeat
+	css_ast.DMaskSize:     cssMaskPrefixTable, // https://caniuse.com/mdn-css_properties_mask-size
+
+	// https://caniuse.com/css-sticky
+	css_ast.DPosition: {
+		IOS:    {prefix: WebkitPrefix, withoutPrefix: v{13, 0, 0}},
+		Safari: {prefix: WebkitPrefix, withoutPrefix: v{13, 0, 0}},
+	},
+
+	// https://caniuse.com/css-color-adjust
+	css_ast.DPrintColorAdjust: {
+		Chrome: {prefix: WebkitPrefix},
+		Edge:   {prefix: WebkitPrefix},
+		Opera:  {prefix: WebkitPrefix},
+		Safari: {prefix: WebkitPrefix, withoutPrefix: v{15, 4, 0}},
+	},
+
+	// https://caniuse.com/css3-tabsize
+	css_ast.DTabSize: {
+		Firefox: {prefix: MozPrefix, withoutPrefix: v{91, 0, 0}},
+		Opera:   {prefix: OPrefix, withoutPrefix: v{15, 0, 0}},
+	},
+
+	// https://caniuse.com/css-text-orientation
+	css_ast.DTextOrientation: {
+		Safari: {prefix: WebkitPrefix, withoutPrefix: v{14, 0, 0}},
+	},
+
+	// https://caniuse.com/text-size-adjust
+	css_ast.DTextSizeAdjust: {
+		Edge: {prefix: MsPrefix, withoutPrefix: v{79, 0, 0}},
+		IOS:  {prefix: WebkitPrefix},
+	},
+
+	// https://caniuse.com/mdn-css_properties_user-select
+	css_ast.DUserSelect: {
+		Chrome:  {prefix: WebkitPrefix, withoutPrefix: v{54, 0, 0}},
+		Edge:    {prefix: MsPrefix, withoutPrefix: v{79, 0, 0}},
+		Firefox: {prefix: MozPrefix, withoutPrefix: v{69, 0, 0}},
+		IOS:     {prefix: WebkitPrefix},
+		Opera:   {prefix: WebkitPrefix, withoutPrefix: v{41, 0, 0}},
+		Safari:  {prefix: WebkitPrefix},
+		IE:      {prefix: MsPrefix},
+	},
+}
+
+func CSSPrefixData(constraints map[Engine][]int) (entries map[css_ast.D]CSSPrefix) {
+	for property, engines := range cssPrefixTable {
+		prefixes := NoPrefix
+		for engine, version := range constraints {
+			if !engine.IsBrowser() {
+				// Specifying "--target=es2020" shouldn't affect CSS
+				continue
+			}
+			if data, ok := engines[engine]; ok && (data.withoutPrefix == v{} || compareVersions(data.withoutPrefix, version) > 0) {
+				prefixes |= data.prefix
+			}
+		}
+		if prefixes != NoPrefix {
+			if entries == nil {
+				entries = make(map[css_ast.D]CSSPrefix)
+			}
+			entries[property] = prefixes
 		}
 	}
 	return