+++ internal/css_ast/css_decl_table.go
@@ -25,6 +25,8 @@ const (
 	DAnimationName
 	DAnimationPlayState
 	DAnimationTimingFunction
+	DAppearance
+	DBackdropFilter
 	DBackfaceVisibility
 	DBackground
 	DBackgroundAttachment
@@ -177,6 +179,7 @@ const (
 	DHyphens
 	DImageOrientation
 	DImageRendering
+	DInitialLetter
 	DInlineSize
 	DInset
 	DJustifyContent
@@ -207,6 +210,7 @@ const (
 	DMask
 	DMaskComposite
 	DMaskImage
+	DMaskOrigin
 	DMaskPosition
 	DMaskRepeat
 	DMaskSize
@@ -259,6 +263,7 @@ const (
 	DPlaceSelf
 	DPointerEvents
 	DPosition
+	DPrintColorAdjust
 	DQuotes
 	DResize
 	DRight
@@ -299,6 +304,7 @@ const (
 	DTextOverflow
 	DTextRendering
 	DTextShadow
+	DTextSizeAdjust
 	DTextTransform
 	DTextUnderlinePosition
 	DTop
@@ -344,6 +350,8 @@ var KnownDeclarations = map[string]D{
 	"animation-name":              DAnimationName,
 	"animation-play-state":        DAnimationPlayState,
 	"animation-timing-function":   DAnimationTimingFunction,
+	"appearance":                  DAppearance,
+	"backdrop-filter":             DBackdropFilter,
 	"backface-visibility":         DBackfaceVisibility,
 	"background":                  DBackground,
 	"background-attachment":       DBackgroundAttachment,
@@ -496,6 +504,7 @@ var KnownDeclarations = map[string]D{
 	"hyphens":                     DHyphens,
 	"image-orientation":           DImageOrientation,
 	"image-rendering":             DImageRendering,
+	"initial-letter":              DInitialLetter,
 	"inline-size":                 DInlineSize,
 	"inset":                       DInset,
 	"justify-content":             DJustifyContent,
@@ -526,6 +535,7 @@ var KnownDeclarations = map[string]D{
 	"mask":                        DMask,
 	"mask-composite":              DMaskComposite,
 	"mask-image":                  DMaskImage,
+	"mask-origin":                 DMaskOrigin,
 	"mask-position":               DMaskPosition,
 	"mask-repeat":                 DMaskRepeat,
 	"mask-size":                   DMaskSize,
@@ -578,6 +588,7 @@ var KnownDeclarations = map[string]D{
 	"place-self":                  DPlaceSelf,
 	"pointer-events":              DPointerEvents,
 	"position":                    DPosition,
+	"print-color-adjust":          DPrintColorAdjust,
 	"quotes":                      DQuotes,
 	"resize":                      DResize,
 	"right":                       DRight,
@@ -618,6 +629,7 @@ var KnownDeclarations = map[string]D{
 	"text-overflow":               DTextOverflow,
 	"text-rendering":              DTextRendering,
 	"text-shadow":                 DTextShadow,
+	"text-size-adjust":            DTextSizeAdjust,
 	"text-transform":              DTextTransform,
 	"text-underline-position":     DTextUnderlinePosition,
 	"top":                         DTop,